fix: Create answer app to isolate from assessment app and added swagger documentation
