+++ answer/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AnswerConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'answer'
