+++ answer/urls.py
@@ -1,0 +1,13 @@
+from django.urls import path, include
+from rest_framework.routers import DefaultRouter
+
+from .views import AnswerViewSet
+
+# Create a router and register our viewsets with it.
+router = DefaultRouter()
+router.register(r"save-answer", AnswerViewSet, basename="save-answer")
+
+# The API URLs are now determined automatically by the router.
+urlpatterns = [
+    path("", include(router.urls)),
+]
