+++ assessment/views.py
@@ -7,13 +7,15 @@ from rest_framework.decorators import api_view, permission_classes
 from django.shortcuts import get_object_or_404
 from django.db import IntegrityError
 from question.models import Question
-from candidate.permissions import CandidatePermission
 
-from .models import Assessment, AssessmentResult, Answer
-from .serializers import AssessmentSerializer, AssessmentResultSerializer, AnswerSerializer
+from .models import Assessment, AssessmentResult
+from .serializers import AssessmentSerializer, AssessmentResultSerializer
 from .permissions import IsAdminOrReadOnly, IsAdminOrCandidate
 
 from .evaluate_assessment import evaluate_assessment, assessment_response
+
+from answer.models import Answer
+from answer.serializers import AnswerSerializer
 
 
 class AssessmentViewSet(viewsets.ModelViewSet):
@@ -27,12 +29,6 @@ class AssessmentResultViewSet(viewsets.ModelViewSet):
     serializer_class = AssessmentResultSerializer
     http_method_names = ['get', 'retrieve']
     permission_classes = [IsAuthenticated, IsAdminOrCandidate]
-
-
-class AnswerViewSet(viewsets.ModelViewSet):
-    queryset = Answer.objects.all()
-    serializer_class = AnswerSerializer
-    permission_classes = [IsAuthenticated, CandidatePermission]
 
 
 @api_view(["POST"])
@@ -55,36 +51,47 @@ def submit_assessment(request):
     response_data = assessment_response(percentage_score, assessment, assessment_results)
     
     # AssessmentResult instance with the calculated total_score
-    try:
-        AssessmentResult.objects.create(
-            candidate=request.user,
-            assessment=assessment,
-            score=float(f"{percentage_score:.2f}"),
-            result = assessment_results,
-        )
-    except IntegrityError as e:
-        return Response({"message": f"This Assessment Already Attempted by {request.user}","error":str(e)}, status=status.HTTP_400_BAD_REQUEST)
+    # try:
+    #     AssessmentResult.objects.create(
+    #         candidate=request.user,
+    #         assessment=assessment,
+    #         score=float(f"{percentage_score:.2f}"),
+    #         result = assessment_results,
+    #     )
+    # except IntegrityError as e:
+    #     return Response({"message": f"This Assessment Already Attempted by {request.user}","error":str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
     return Response(response_data, status=status.HTTP_201_CREATED)
 
 @api_view(["POST"])
 @permission_classes([IsAdminUser])
-def update_question(request, action):
+def add_question(request, action):
     assessment_id = request.data.get("assessment_id")
     question_id = request.data.get("question_id")
 
     try:
         assessment = Assessment.objects.get(id=assessment_id)
         question = Question.objects.get(id=question_id)
+        assessment.questions.add(question)
+        message = "Question added successfully."
 
-        if action == "add":
-            assessment.questions.add(question)
-            message = "Question added successfully."
-        elif action == "remove":
-            assessment.questions.remove(question)
-            message = "Question removed successfully."
-        else:
-            return Response({"message": "Invalid action."}, status=status.HTTP_400_BAD_REQUEST)
+        return Response({"message": message}, status=status.HTTP_200_OK)
+    except Assessment.DoesNotExist:
+        return Response({"message": "Assessment not found."}, status=status.HTTP_404_NOT_FOUND)
+    except Question.DoesNotExist:
+        return Response({"message": "Question not found."}, status=status.HTTP_404_NOT_FOUND)
+
+@api_view(["POST"])
+@permission_classes([IsAdminUser])
+def remove_question(request, action):
+    assessment_id = request.data.get("assessment_id")
+    question_id = request.data.get("question_id")
+
+    try:
+        assessment = Assessment.objects.get(id=assessment_id)
+        question = Question.objects.get(id=question_id)
+        assessment.questions.remove(question)
+        message = "Question removed successfully."
 
         return Response({"message": message}, status=status.HTTP_200_OK)
     except Assessment.DoesNotExist:
